// src/sappurchaseinvoices/application/interfaces/service-layer-purchase-invoice.interface.ts
export interface ServiceLayerPurchaseInvoiceLine {
  readonly ItemCode?: string;
  readonly ItemDescription?: string;
  readonly Quantity: number;
  readonly UnitPrice: number;
  readonly LineTotal: number;
  readonly TaxCode?: string;
  readonly TaxLiable?: 'tYES' | 'tNO';
  readonly AccountCode: string;
  readonly CostingCode?: string;
  readonly LineMemo: string;
  readonly Text?: string;
  readonly LineText?: string;
  readonly U_RS_UUID?: string;
  readonly U_RS_RFC?: string;
  readonly U_SelloCFDI?: string;
  readonly U_SelloSAT?: string;
  readonly U_CertiSAT?: string;
  readonly U_SAT_SERIECFD?: string;
  readonly U_FECHAEMISION?: string;
  readonly U_CFDI4_EXPORTACION?: string;
  readonly U_CFDI4_OBJETOIMP?: string;
  readonly U_SAT_METPAGO?: string;
  readonly U_SAT_FOLIO?: string;
  readonly U_FAE_NUM_CTA?: string;
  readonly U_FAE_MET_PAGO?: string;
  readonly U_CadenaCFDI?: string;
  readonly U_FECHATIMBRE?: string;
  readonly U_UDF_UUID?: string;
}

export interface ServiceLayerPurchaseInvoicePayload {
  readonly CardCode: string;
  readonly DocDate: string;
  readonly DocDueDate: string;
  readonly TaxDate: string;
  readonly DocTotal: number;
  readonly DocTotalFC: number;
  readonly Comments: string;
  readonly JournalMemo: string;
  readonly DocumentLines: readonly ServiceLayerPurchaseInvoiceLine[];
  readonly DocType?: string;
  readonly Reference1?: string;
  readonly Reference2?: string;
  readonly VatSum?: number;
  readonly TotalDiscount?: number;
  readonly DocCurrency?: string;
  readonly PaymentMethod?: string;
  readonly PaymentTerms?: string;
  readonly ShipToCode?: string;
  readonly FederalTaxID?: string;
  readonly U_XML?: string;
  readonly U_PDF?: string;
  readonly U_RS_UUID?: string;
  readonly U_RS_RFC?: string;
  readonly U_UUID?: string;
  readonly U_SAT_SERIECFD?: string;
  readonly U_FECHAEMISION?: string;
  readonly U_CFDI4_EXPORTACION?: string;
  readonly U_FAE_NUM_CTA?: string;
  readonly U_FAE_MET_PAGO?: string;
  readonly U_SelloSAT?: string;
  readonly U_SelloCFDI?: string;
  readonly U_CertiSAT?: string;
  readonly U_CadenaCFDI?: string;
  readonly U_FECHATIMBRE?: string;
  readonly NumAtCard?: string;
}